Examples details added to README. Simple example updated to have todos.
