+++ examples/simple/config/babel.prod.js
@@ -10,7 +10,7 @@ module.exports = {
     'babel-plugin-transform-class-properties',
     'babel-plugin-transform-object-rest-spread',
     'babel-plugin-transform-react-constant-elements',
-    'babel-plugin-transform-decorators-legacy',
+    'babel-plugin-transform-decorators-legacy'
   ].map(require.resolve).concat([
     [require.resolve('babel-plugin-transform-runtime'), {
       helpers: false,
@@ -18,4 +18,4 @@ module.exports = {
       regenerator: true
     }]
   ])
-};
+}
