+++ examples/simple/package.json
@@ -45,6 +45,7 @@
     "whatwg-fetch": "1.0.0"
   },
   "dependencies": {
+    "lodash": "^4.15.0",
     "react": "^15.2.1",
     "react-dom": "^15.2.1",
     "react-redux": "^4.4.5",
