+++ examples/simple/src/Todo.css
@@ -1,0 +1,11 @@
+.Todo {
+  margin: 1rem;
+
+}
+.Todo-Input {
+  margin: 1rem;
+}
+
+.Todo-Button {
+  margin: 1rem;
+}
